sample-app/app/producer/requirements.txt
pylint==2.4.0
Flask==1.1.2
kafka-python==2.0.1
gunicorn==20.0.4
cloudevents==0.2.4
requests==2.23.0
